+++ e2e-apps/github-issues/src/issue-fixtures/issue-999/pages/page/index.ts
@@ -1,0 +1,6 @@
+/* eslint-disable ts/no-require-imports */
+require('../../subs/page/callback.ts', () => {}, () => {})
+const promise = require.async('../../subs/page/promise.ts')
+const native = import('../../subs/page/native.ts')
+
+Page({ promise, native })