+++ e2e-apps/wot-ui-compat/CHANGELOG.md
@@ -5,6 +5,20 @@
 ### Patch Changes
 
 - Updated dependencies:
+  - wevu@7.1.2
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies:
+  - wevu@7.1.1
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies:
   - wevu@7.1.0
 
 ## 0.0.9