+++ templates/weapp-vite-lib-template/CHANGELOG.md
@@ -5,6 +5,20 @@
 ### Patch Changes
 
 - Updated dependencies:
+  - wevu@7.1.2
+
+## 11.0.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - wevu@7.1.1
+
+## 11.0.3
+
+### Patch Changes
+
+- Updated dependencies:
   - wevu@7.1.0
 
 ## 11.0.3